Refactoring(server-info): all subsystem-specific parsers moved to appropriate subsystem class.
