+++ netutils_linux_hardware/memory.py
@@ -1,6 +1,9 @@
 # coding=utf-8
+import yaml
+from six import iteritems
 
 from netutils_linux_hardware.grade import Grade
+from netutils_linux_hardware.parser import YAMLLike, Parser
 from netutils_linux_hardware.subsystem import Subsystem
 
 
@@ -42,3 +45,57 @@ class Memory(Subsystem):
             'MemTotal': Grade.int(size.get('MemTotal'), 2 * (1024 ** 2), 16 * (1024 ** 2)),
             'SwapTotal': Grade.int(size.get('SwapTotal'), 512 * 1024, 4 * (1024 ** 2)),
         }, self.folding.DEVICE) if size else 1
+
+
+class MemInfo(YAMLLike):
+    keys_required = (
+        'MemTotal',
+        'MemFree',
+        'SwapTotal',
+        'SwapFree',
+    )
+
+    def parse(self, text):
+        return dict((k, int(v.replace(' kB', ''))) for k, v in iteritems(yaml.load(text)) if k in self.keys_required)
+
+
+class MemInfoDMIDevice(object):
+    def __init__(self, text):
+        self.data = {
+            'speed': 0,
+            'type': 'RAM',
+            'size': 0,
+        }
+        self.handle = None
+        self.parse_text(text)
+
+    def parse_text(self, text):
+        """ Разбор описания плашки памяти от dmidecode """
+        for line in map(str.strip, text.split('\n')):
+            self.parse_line(line)
+
+    def parse_line(self, line):
+        for key in ('Speed', 'Type', 'Size'):
+            if line.startswith(key + ':'):
+                self.data[key.lower()] = line.split()[1]
+                break
+        if line.startswith('Handle'):
+            self.handle = line.split(' ')[1].strip(',')
+
+
+class MemInfoDMI(Parser):
+    @staticmethod
+    def parse(text):
+        """ Разбор всего вывода dmidecode --type memory """
+        return MemInfoDMI.__parse(text.split('\n\n')) if text else None
+
+    @staticmethod
+    def __parse(devices):
+        output = dict()
+        for device in devices:
+            if 'Memory Device' in device:
+                mem_dev = MemInfoDMIDevice(device)
+                if mem_dev.data.get('size') == 'No':
+                    continue
+                output[mem_dev.handle] = mem_dev.data
+        return output
