+++ netutils_linux_hardware/net.py
@@ -1,6 +1,10 @@
 # coding=utf-8
+import os
+
+from six import iteritems, print_
 
 from netutils_linux_hardware.grade import Grade
+from netutils_linux_hardware.parser import Parser
 from netutils_linux_hardware.rater_math import extract
 from netutils_linux_hardware.subsystem import Subsystem
 
@@ -37,3 +41,72 @@ class Net(Subsystem):
                 'max': Grade.int(buffers.get('max'), 256, 4096),
             }, self.folding.DEVICE)
         }, self.folding.DEVICE)
+
+
+class ReductorMirror(Parser):
+    @staticmethod
+    def parse(text):
+        lines = dict((line.split(' ', 1)) for line in text.strip().split('\n'))
+        for netdev, conf in iteritems(lines):
+            output = dict()
+            output['conf'] = dict()
+            output['conf']['vlan'], output['conf']['ip'] = conf.split()
+            output['conf']['vlan'] = output['conf']['vlan'] != '-'
+            if output['conf']['ip'] == '-':
+                output['conf']['ip'] = ''
+            lines[netdev] = output
+        return lines
+
+
+class NetdevParser(Parser):
+    @staticmethod
+    def parse(netdev_keys):
+        """
+        :param netdev_keys: list of device names
+        :return: dict[device_name] = {'vlan': bool; 'ip': ''}
+        """
+        netdevs = dict()
+        for key in netdev_keys:
+            if key.count('.') == 1:
+                dev, _ = key.split('.')
+            elif key.count('.') == 0:
+                dev = key
+            else:
+                print_('QinQ not supported yet. Device: {0}'.format(key))
+                raise NotImplementedError
+            netdevs[dev] = dict()
+            netdevs[dev]['conf'] = {
+                'vlan': key.count('.') == 1,
+                'ip': '',
+            }
+        return netdevs
+
+
+class BridgeOutput(NetdevParser):
+    @staticmethod
+    def parse(text):
+        """ # 0 - id, 1 - dev, 2 - _, 3 - state, 4 - _, 5 - details, 6 - _, 7 - mtu, 8 - _, 9 - master
+        :param text: # 3: eth1 state DOWN : <NO-CARRIER,BROADCAST,MULTICAST,UP> mtu 1500 master br1 state disabled
+        :return: analyzed netdevs
+        """
+        __dev__ = 1
+        netdevs_keys = [line.split()[__dev__] for line in text.strip().split('\n')]
+        return NetdevParser.parse(netdevs_keys)
+
+
+class EthtoolFiles(NetdevParser):
+    def parse_file(self, filepath, **kwargs):
+        return self.parse(os.listdir(filepath))
+
+
+class EthtoolBuffers(Parser):
+    @staticmethod
+    def parse(text):
+        buffers = [int(line.split()[1])
+                   for line in text.strip().split('\n')
+                   if line.startswith('RX:')]
+        if buffers and len(buffers) == 2:
+            return {
+                'max': buffers[0],
+                'cur': buffers[1],
+            }
