+++ netutils_linux_hardware/test_parsers.py
@@ -3,7 +3,7 @@
 
 from unittest import TestCase
 
-from netutils_linux_hardware.parser import DiskInfo
+from netutils_linux_hardware.disk import DiskInfo
 
 
 class DiskInfoTests(TestCase):
